feat(ui/discovery): suggestion for the search box
